Russian translation is fully completed (#358)
* Create ru_ru

* Delete ru_ru

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* Added ru_ru macros

* Updated ru_ru lang

*Fixed a lot of translation errors that created misunderstandings.

* Added translation for the last added items like marshmallow or redstone tools.

* The hints in JEI and Waila have become much more understandable.

* Added entry for Russian players.

It will help solve the font trouble that many Russian players may face. The entry has to be added to all versions of the translation due to the peculiarities of Patchouli mod.

* Added entry for Russian players

It will help solve the font trouble that many Russian players may face. The entry has to be added to all versions of the translation due to the peculiarities of Patchouli mod.

* Added full ru_ru Patchouli translation

* Update ru_ru.lang

Minor adjustments to outdated information.

* Minor fixes, last.

Excuse me. I'm too anxious.

* Update ru_ru.lang

* Added lost line